[refactor] Consolidate ResolvedConstructor creation logic
