+++ src/FakeItEasy/Creation/ResolvedConstructor.cs
@@ -1,19 +1,52 @@
 namespace FakeItEasy.Creation
 {
+    using System;
+    using System.Collections.Generic;
     using System.Linq;
 
     internal class ResolvedConstructor
     {
-        public ResolvedConstructor(ResolvedArgument[] resolvedArguments)
+        public ResolvedConstructor(
+            IEnumerable<Type> parameterTypes,
+            DummyCreationSession session,
+            IDummyValueResolver resolver)
         {
-            Guard.AgainstNull(resolvedArguments, nameof(resolvedArguments));
-            this.Arguments = resolvedArguments;
+            this.Arguments = ResolveArguments(parameterTypes, session, resolver);
         }
 
         public bool WasSuccessfullyResolved => this.Arguments.All(x => x.WasResolved);
 
-        public ResolvedArgument[] Arguments { get; }
+        public IEnumerable<ResolvedArgument> Arguments { get; }
 
         public string ReasonForFailure { get; set; }
+
+        private static IEnumerable<ResolvedArgument> ResolveArguments(
+            IEnumerable<Type> parameterTypes,
+            DummyCreationSession session,
+            IDummyValueResolver resolver)
+        {
+            var resolvedArguments = new List<ResolvedArgument>();
+            foreach (var parameterType in parameterTypes)
+            {
+                var resolvedArgument = new ResolvedArgument { ArgumentType = parameterType };
+                try
+                {
+                    var creationResult = resolver.TryResolveDummyValue(session, parameterType);
+                    resolvedArgument.WasResolved = creationResult.WasSuccessful;
+                    if (creationResult.WasSuccessful)
+                    {
+                        resolvedArgument.ResolvedValue = creationResult.GetResultAsDummy();
+                    }
+                }
+                catch
+                {
+                    resolvedArgument.WasResolved = false;
+                }
+
+                resolvedArguments.Add(resolvedArgument);
+            }
+
+            return resolvedArguments;
+        }
     }
 }
